+++ histore/document/csv/read.py
@@ -7,16 +7,24 @@
 
 """Functions and classes to read CSV files as HISTORE documents."""
 
-import csv
-import gzip
-import pandas as pd
+from typing import List, Optional
 
-from histore.document.base import Document
+from histore.document.base import Document, PrimaryKey
+from histore.document.csv.base import CSVFile, CSVReader
+from histore.document.csv.sort import SortedCSVDocument
+from histore.document.mem.base import InMemoryDocument
 from histore.document.mem.dataframe import DataFrameDocument
 from histore.document.reader import DocumentReader
 from histore.document.row import DocumentRow
+from histore.document.schema import Column
 from histore.key.base import NumberKey
 
+import histore.document.csv.sort as sort
+import histore.document.schema as schema
+import histore.key.annotate as anno
+
+
+# -- Documents ----------------------------------------------------------------
 
 class SimpleCSVDocument(Document):
     """CSV file containing a document that is keyed by the row position and
@@ -24,43 +32,16 @@ class SimpleCSVDocument(Document):
     document order with the row identifier being the line number (starting at
     zero and ignoring an optional column header).
     """
-    def __init__(
-        self, filename, columns, delimiter=',', quotechar='"',
-        compression=None, has_header=True
-    ):
-        """Initialize the object properties. The abstract document class
-        maintains a list of column names in the document schema. Columns may
-        either be represented by strings or by instances of the Column class.
+    def __init__(self, file: CSVFile):
+        """Initialize the reference to the CSV file reader.
 
         Parameters
         ----------
-        filename: string
-            Path to the CSV file that contains the document.
-        columns: list
-            List of column names. The number of values in each document row is
-            expected to be the same as the number of columns and the order of
-            values in each row is expected to correspond to their respective
-            column in this list.
-        delimiter: string, default=','
-            A one-character string used to separate fields.
-        quotechar: string, default='"'
-            A one-character string used to quote fields containing special
-            characters, such as the delimiter or quotechar, or which contain
-            new-line characters.
-        compression: string, default=None
-            String identifier for the file compression format. Currently only
-            'gzip' compression is supported.
-        has_header: bool, default=True
-            Flag indicating whether the first row of the given file contains
-            column names. The header row will be ignored by the reader in favor
-            of the given column names.
+        file: histore.document.csv.base.CSVFile
+            Reference to the CSV file.
         """
-        self.filename = filename
-        self.columns = columns
-        self.delimiter = delimiter
-        self.quotechar = quotechar
-        self.compression = compression
-        self.has_header = has_header
+        super(SimpleCSVDocument, self).__init__(columns=file.columns)
+        self.file = file
 
     def close(self):
         """The un-keyed CSV document reader has no additional resources that
@@ -68,22 +49,10 @@ class SimpleCSVDocument(Document):
         """
         pass  # pragma: no cover
 
-    def open(self):
-        """Open the associated file for reading.
-
-        Returns
-        -------
-        FileObject
-        """
-        if self.compression == 'gzip':
-            return gzip.open(self.filename, 'rt')
-        else:
-            return open(self.filename, 'r')
-
     def partial(self, reader):
         """If the document in the CSV file is a partial document we read it
-        into a data frame and use the respective data frame document method.
-        This requires that partial documents in CSV files fit into main memory.
+        as a data frame. This requires that partial documents in CSV files fit
+        into main memory.
 
         Parameters
         ----------
@@ -95,17 +64,9 @@ class SimpleCSVDocument(Document):
         -------
         histore.document.base.Document
         """
-        df = pd.read_csv(
-            self.filename,
-            delimiter=self.delimiter,
-            quotechar=self.quotechar,
-            header=0 if self.has_header else None,
-            names=self.columns,
-            compression=self.compression
-        )
-        return DataFrameDocument(df).partial(reader)
+        return DataFrameDocument(self.file.read_df()).partial(reader)
 
-    def reader(self, schema):
+    def reader(self, schema: List[Column]):
         """Get reader for data frame rows ordered by their row identifier. In
         a partial document the row positions that are returned by the reader
         are aligned with the positions of the corresponding rows in the
@@ -130,32 +91,17 @@ class SimpleCSVDocument(Document):
         ------
         ValueError
         """
-        # Open the input file and skip the header.
-        f = self.open()
-        try:
-            reader = csv.reader(
-                f,
-                delimiter=self.delimiter,
-                quotechar=self.quotechar
-            )
-            if self.has_header:
-                # Skip the column header (if present).
-                next(reader)
-        except IOError as ex:  # pragma: no cover
-            # Ensure that the file is closed in case of an error.
-            f.close()
-            raise ex
-        return SimpleCSVDocumentReader(reader=reader, schema=schema, fh=f)
+        return SimpleCSVDocumentReader(reader=self.file.open(), schema=schema)
 
 
 class SimpleCSVDocumentReader(DocumentReader):
     """Document reader for un-keyed CSV documents."""
-    def __init__(self, reader, schema, fh):
+    def __init__(self, reader: CSVReader, schema: List[Column]):
         """Initialize the CSV reader and the list of column names.
 
         Parameters
         ----------
-        reader: csv.reader
+        reader: histore.document.csv.base.CSVReader
             Reader for the document file. If the file has a header it is
             assumed that the corresponding row has already been read.
         schema: list(histore.document.schema.Column)
@@ -163,15 +109,10 @@ class SimpleCSVDocumentReader(DocumentReader):
             a column in the document rows (based on list position). The schema
             columns provide the unique column identifier that are required to
             generate the returned document rows.
-        fh: FileObject
-            Handle for the file that is being read. Required to be able to
-            properly close the file when done with reading.
         """
         self.reader = reader
         self.schema = schema
-        self.fh = fh
         self._next_row = None
-        self._row_id = 0
         self.next()
 
     def has_next(self):  # pragma: no cover
@@ -194,14 +135,13 @@ class SimpleCSVDocumentReader(DocumentReader):
         """
         result = self._next_row
         try:
-            row = next(self.reader)
+            rowid, row = next(self.reader)
             # If a next row was read ensure that the number of values
             # is the same as the number of columns in the schema.
             if len(row) != len(self.schema):
-                self.fh.close()
-                raise ValueError('invalid row %d' % (self.reader.line_num))
+                self.reader.close()
+                raise ValueError('invalid row %d' % (self.reader.rowid))
             # Create document row object for the read values.
-            rowid = self._row_id
             values = dict()
             for i, col in enumerate(self.schema):
                 values[col.colid] = row[i]
@@ -210,10 +150,82 @@ class SimpleCSVDocumentReader(DocumentReader):
                 pos=rowid,
                 values=values
             )
-            self._row_id += 1
         except StopIteration:
             # Close the file if the file is reached.
             self._next_row = None
-            self.fh.close()
+            self.reader.close()
         # Return the buffered result
         return result
+
+
+def open_document(
+    file: CSVFile, primary_key: Optional[PrimaryKey] = None,
+    max_size: Optional[float] = None
+) -> Document:
+    """Read a document from a csv file. If the primary key attributes are
+    given the returned document is sorted by the primary key values.
+    Otherwise, the original order of rows is kept and each row is assigned
+    a unique index that is equal to the row position. Sorts the document by
+    the primary key values (if given).
+
+    Parameters
+    ----------
+    file: histore.document.csv.base.CSVFile
+        Reference to the CSV file.
+    primary_key: string or list, default=None
+        Column(s) that are used to generate identifier for snapshot rows.
+        The columns may be identified by their name or index position. If a
+        string in the primary key list refers to a non-unique column name
+        in the file, a ValueError is raised.
+    max_size: float, default=None
+        Maximum size (in MB) of the main-memory buffer for blocks of the
+        CSV file that are sorted in main-memory.
+
+    Returns
+    -------
+    histore.document.base.Document
+
+    Raises
+    ------
+    ValueError
+    """
+    # with self.open() as reader:
+    # Create document instance depending on whether a primary key was given
+    # or not.
+    if primary_key is not None:
+        # If a primary key is given we first need to get the index position
+        # for the key attributes in the document schema and then sort the
+        # input file.
+        columns = file.columns
+        pk = schema.column_index(schema=columns, columns=primary_key)
+        with file.open() as reader:
+            buffer, filenames = sort.split(
+                reader=reader,
+                sortkey=pk,
+                buffer_size=max_size
+            )
+        if not filenames:
+            # If the file fits into main-memory return a sorted in-memory
+            # document.
+            return InMemoryDocument(
+                columns=columns,
+                rows=buffer,
+                readorder=anno.pk_readorder(rows=buffer, primary_key=pk)
+            )
+        else:
+            # Merge the CSV file blocks and return a sorted CSV document
+            # object that wrapps the sorted CSV file.
+            return SortedCSVDocument(
+                filename=sort.mergesort(
+                    buffer=buffer,
+                    filenames=filenames,
+                    sortkey=pk
+                ),
+                columns=columns,
+                primary_key=pk
+            )
+    else:
+        # In this case we do not need to sort the document. The document
+        # will always be read in original order. Return a file reader for
+        # the csv file.
+        return SimpleCSVDocument(file)